+++ src/arch/amd64/include/sys/arch/space/address_space.hh
@@ -6,6 +6,7 @@ namespace sys::arch::space
     struct address_space {};
     inline void initialize(address_space&, u16, word_t) noexcept {}
     inline void activate(address_space&) noexcept {}
+    inline void activate_kernel() noexcept {}
     inline void invalidate_asid(u16) noexcept {}
     inline error_t map_page(address_space&, vaddr_t, void*, bool, bool) noexcept {
         return error_t::unsupported;